+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,57 +1,5 @@
 import NextAuth from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { NextAuthOptions } from "next-auth";
-
-const users = [
-  {
-    id: "1",
-    email: "admin@example.com",
-    password: "admin123",
-    name: "Admin User",
-    role: "admin",
-  },
-];
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error("Please enter an email and password");
-        }
-
-        const user = users.find((user) => user.email === credentials.email);
-
-        if (!user || user.password !== credentials.password) {
-          throw new Error("Invalid email or password");
-        }
-          
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const { password, ...userWithoutPassword } = user;
-        return userWithoutPassword;
-      },
-    }),
-  ],
-  pages: {
-    signIn: "/signin",
-  },
-  callbacks: {
-    async jwt({ token }) {
-      return token;
-    },
-    async session({ session }) {
-      return session;
-    },
-  },
-  session: {
-    strategy: "jwt",
-  },
-};
+import { authOptions } from "@/lib/auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
